Make current_context() return None instead of exception
If no context is set, we'll return None instead of an exception.
I'd like to reserve exceptions for truly exceptional cases. I
think the current-context kubectl command exit(1)'s as a sane but
noisy default for shell environments that have strange type
systems.

In Python land, the None fits the bill better than an exception
in this particular case.